Add test suite for decorators
