+++ modules/person.rb
@@ -1,8 +1,4 @@
-class Nameable
-  def correct_name
-    raise NotImplementedError
-  end
-end
+require_relative './decorator'
 
 class Person < Nameable
   attr_accessor :name, :age, :parent_permission
@@ -34,28 +30,5 @@ class Person < Nameable
 
   def of_age?
     @age >= 18
-  end
-end
-
-class BaseDecorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @namealbe
-  end
-end
-
-class CapitalizeDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.upcase
-  end
-end
-
-class TrimmerDecorator < BaseDecorator
-  def correct_name
-    @nameable.correct_name.slice(0, 10)
   end
 end
