+++ spec/spec_helper.rb
@@ -1,10 +1,10 @@
+require './modules/decorator'
 require './modules/student'
 require './modules/classroom'
 require './modules/book'
 require './modules/person'
 require './modules/teacher'
 require './modules/rental'
-
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
